pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "emperator"
version = "0.1.0"
description = "Canonical implementation scaffold for the Emperator platform."
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Emperator Contributors" }]
license = { text = "Apache-2.0" }
dependencies = [
  "fastapi ~= 0.119",
  "uvicorn[standard] ~= 0.37",
  "pyyaml ~= 6.0",
  "typer ~= 0.19",
  "rich ~= 14.0",
  "shellingham ~= 1.5",
  "types-PyYAML ~= 6.0",
  "tree-sitter ~= 0.23",
  "tree-sitter-python ~= 0.23",
  "tree-sitter-javascript ~= 0.23",
  "msgpack ~= 1.1",
  "libcst ~= 1.5",
]

[project.optional-dependencies]
dev = [
  "pytest ~= 8.4",
  "pytest-cov ~= 7.0",
  "ruff ~= 0.7.0",
  "mypy ~= 1.18",
  "httpx ~= 0.28",
  "pre-commit ~= 4.0",
  "mdformat ~= 0.7",
  "mdformat-gfm ~= 0.3",
  "bandit ~= 1.7",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
fix = true
extend-exclude = [
  "site",
  "node_modules",
  ".venv",
  ".mypy_cache",
  ".ruff_cache",
  ".pytest_cache",
  "**/._*",
]

[tool.ruff.lint]
select = ["ALL"]
ignore = [
  "ANN101",  # self annotations unnecessary for Typer callbacks
  "ANN102",  # cls annotations unnecessary for Typer callbacks
  "COM812",  # avoid conflicts with formatter managed commas
  "D100",    # module docstrings redundant in tests
  "D101",
  "D102",
  "D103",
  "D104",
  "D105",
  "D106",
  "D107",
  "D203",
  "D213",
  "ISC001",  # prefer formatting plugin to manage implicit string concatenation
  "N818",    # pytest fixtures use mixed casing
  "PLR2004", # allow magic values in tests
  "S101",    # allow assert in tests
  "TCH001",  # allow direct typing imports for clarity
  "TCH002",
  "TCH003",
  "Q000",    # Ruff formatter enforces single quotes, disable conflicting lint
  "Q003",
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = [
  "ANN201",
  "ANN401",
  "ANN001",
  "ANN002",
  "ANN003",
  "ANN202",
  "S311",
  "S603",
  "S607",
  "SIM117",
  "ARG001",
  "ARG005",
  "FBT001",
  "FBT002",
  "FBT003",
  "PT018",
  "TRY003",
  "EM101",
  "EM102",
  "INP001",
  "SLF001",
]
"scripts/*.mjs" = ["INP001"]

[tool.ruff.format]
quote-style = "single"

[tool.pytest.ini_options]
addopts = "-q -ra --strict-markers"

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
ignore_missing_imports = true

[tool.bandit]
skips = ["B101"]

[project.scripts]
emperator = "emperator.cli:run"
